+++ packages/shared/src/index.server.ts
@@ -36,6 +36,8 @@ export * from "./constants.js";
 export {
     resolveEnvironmentVariableOverridesFromConfig,
     loadConfig,
+    getIdentityProviderConfigs,
+    getIdentityProviderConfig,
     isRemotePath,
 } from "./env.server.js";
 export { env } from "./env.server.js"